Make 'completed' prop in TodoList optional
Relaxed the requirement of the 'completed' prop in both TodoList and TodoListItem components from being mandatory to optional. This adjustment accommodates todo items that may not have a completed status initially defined, aiding in the integration with data sources that don't necessarily provide this attribute. This change increases the flexibility of todo item handling.